requirements.txt
django==3.1.6
python-dotenv
django-sass==1.0.0
fontawesome-free
Pillow
opencv-python
requests